backend/requirements.txt
flask
flask-cors
python-dotenv
opencv-python==4.10.0.84
numpy==2.0.2
google-genai>=0.3.0
pytest
pytest-flask
